+++ CHANGELOG.rst
@@ -6,7 +6,7 @@ Version 0.2
 
 - Basic async client
 - Export design physical properties
-- Fusion360 server now runs in a worker thread
+- Fusion360 server can now run in a worker thread
 - Expose a function to batch commands that preserves the call order
 
 Version 0.1